use 10bit values from hsv conversion directly for pwm module

+++ software/color.h
@@ -4,10 +4,12 @@
 #include <stdint.h>
 
 struct color_rgb {
-  uint8_t r, g, b;
+  // Values from 0 .. 1023.
+  uint16_t r, g, b;
 };
 
 struct color_hsv {
+  // h from 0 .. 1535; s and v 0 .. 255.
   uint16_t h;
   uint8_t s, v;
 };
